fs 동기 작업
Wait, I thought using the node.js' synchronous APIs was bad form?

We're not writing a server so there's not nearly the same IO demand on the process

Also if we used an async API and our app was quit before the asynchronous write had a chance to complete,

we might lose that data. Note that in a real app, we would try/catch this.